consolidate store_utils into store
